자바 ORM 표준 JPA 프로그래밍 - 기본편 연관관계 매핑 기초 - 단방향 연관관계 (데이터 중심 모델링)
