+++ jpa-basic/src/main/java/jpabook/jpashop/domain/Item.java
@@ -1,56 +1,56 @@
-package jpabook.jpashop.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-public class Item {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "ITEM_ID")
-    private Long id;
-
-    private String name;
-
-    private int price;
-
-    private int stockQuantity;
-
-    protected Item() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public void setPrice(int price) {
-        this.price = price;
-    }
-
-    public int getStockQuantity() {
-        return stockQuantity;
-    }
-
-    public void setStockQuantity(int stockQuantity) {
-        this.stockQuantity = stockQuantity;
-    }
-}
+//package jpabook.jpashop.domain;
+//
+//import javax.persistence.Column;
+//import javax.persistence.Entity;
+//import javax.persistence.GeneratedValue;
+//import javax.persistence.Id;
+//
+//@Entity
+//public class Item {
+//
+//    @Id
+//    @GeneratedValue
+//    @Column(name = "ITEM_ID")
+//    private Long id;
+//
+//    private String name;
+//
+//    private int price;
+//
+//    private int stockQuantity;
+//
+//    protected Item() {
+//    }
+//
+//    public Long getId() {
+//        return id;
+//    }
+//
+//    public void setId(Long id) {
+//        this.id = id;
+//    }
+//
+//    public String getName() {
+//        return name;
+//    }
+//
+//    public void setName(String name) {
+//        this.name = name;
+//    }
+//
+//    public int getPrice() {
+//        return price;
+//    }
+//
+//    public void setPrice(int price) {
+//        this.price = price;
+//    }
+//
+//    public int getStockQuantity() {
+//        return stockQuantity;
+//    }
+//
+//    public void setStockQuantity(int stockQuantity) {
+//        this.stockQuantity = stockQuantity;
+//    }
+//}
