+++ jpa-basic/src/main/java/jpabook/jpashop/domain/Order.java
@@ -1,63 +1,63 @@
-package jpabook.jpashop.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "ORDERS")
-public class Order {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "ORDER_ID")
-    private Long id;
-
-    @Column(name = "MEMBER_ID")
-    private Long memberId;
-
-    private LocalDateTime orderDate;
-
-    @Enumerated(EnumType.STRING)
-    private OrderStatus status;
-
-    protected Order() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getMemberId() {
-        return memberId;
-    }
-
-    public void setMemberId(Long memberId) {
-        this.memberId = memberId;
-    }
-
-    public LocalDateTime getOrderDate() {
-        return orderDate;
-    }
-
-    public void setOrderDate(LocalDateTime orderDate) {
-        this.orderDate = orderDate;
-    }
-
-    public OrderStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(OrderStatus status) {
-        this.status = status;
-    }
-}
+//package jpabook.jpashop.domain;
+//
+//import javax.persistence.Column;
+//import javax.persistence.Entity;
+//import javax.persistence.EnumType;
+//import javax.persistence.Enumerated;
+//import javax.persistence.GeneratedValue;
+//import javax.persistence.Id;
+//import javax.persistence.Table;
+//import java.time.LocalDateTime;
+//
+//@Entity
+//@Table(name = "ORDERS")
+//public class Order {
+//
+//    @Id
+//    @GeneratedValue
+//    @Column(name = "ORDER_ID")
+//    private Long id;
+//
+//    @Column(name = "MEMBER_ID")
+//    private Long memberId;
+//
+//    private LocalDateTime orderDate;
+//
+//    @Enumerated(EnumType.STRING)
+//    private OrderStatus status;
+//
+//    protected Order() {
+//    }
+//
+//    public Long getId() {
+//        return id;
+//    }
+//
+//    public void setId(Long id) {
+//        this.id = id;
+//    }
+//
+//    public Long getMemberId() {
+//        return memberId;
+//    }
+//
+//    public void setMemberId(Long memberId) {
+//        this.memberId = memberId;
+//    }
+//
+//    public LocalDateTime getOrderDate() {
+//        return orderDate;
+//    }
+//
+//    public void setOrderDate(LocalDateTime orderDate) {
+//        this.orderDate = orderDate;
+//    }
+//
+//    public OrderStatus getStatus() {
+//        return status;
+//    }
+//
+//    public void setStatus(OrderStatus status) {
+//        this.status = status;
+//    }
+//}
