+++ jpa-basic/src/main/java/jpabook/jpashop/domain/OrderItem.java
@@ -1,68 +1,68 @@
-package jpabook.jpashop.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "ORDER_ITEM")
-public class OrderItem {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "ORDER_ITEM_ID")
-    private Long id;
-
-    @Column(name = "ORDER_ID")
-    private Long orderId;
-    @Column(name = "ITEM_ID")
-    private Long itemId;
-
-    private int orderPrice;
-    private int count;
-
-    public OrderItem() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(Long orderId) {
-        this.orderId = orderId;
-    }
-
-    public Long getItemId() {
-        return itemId;
-    }
-
-    public void setItemId(Long itemId) {
-        this.itemId = itemId;
-    }
-
-    public int getOrderPrice() {
-        return orderPrice;
-    }
-
-    public void setOrderPrice(int orderPrice) {
-        this.orderPrice = orderPrice;
-    }
-
-    public int getCount() {
-        return count;
-    }
-
-    public void setCount(int count) {
-        this.count = count;
-    }
-}
+//package jpabook.jpashop.domain;
+//
+//import javax.persistence.Column;
+//import javax.persistence.Entity;
+//import javax.persistence.GeneratedValue;
+//import javax.persistence.Id;
+//import javax.persistence.Table;
+//
+//@Entity
+//@Table(name = "ORDER_ITEM")
+//public class OrderItem {
+//
+//    @Id
+//    @GeneratedValue
+//    @Column(name = "ORDER_ITEM_ID")
+//    private Long id;
+//
+//    @Column(name = "ORDER_ID")
+//    private Long orderId;
+//    @Column(name = "ITEM_ID")
+//    private Long itemId;
+//
+//    private int orderPrice;
+//    private int count;
+//
+//    public OrderItem() {
+//    }
+//
+//    public Long getId() {
+//        return id;
+//    }
+//
+//    public void setId(Long id) {
+//        this.id = id;
+//    }
+//
+//    public Long getOrderId() {
+//        return orderId;
+//    }
+//
+//    public void setOrderId(Long orderId) {
+//        this.orderId = orderId;
+//    }
+//
+//    public Long getItemId() {
+//        return itemId;
+//    }
+//
+//    public void setItemId(Long itemId) {
+//        this.itemId = itemId;
+//    }
+//
+//    public int getOrderPrice() {
+//        return orderPrice;
+//    }
+//
+//    public void setOrderPrice(int orderPrice) {
+//        this.orderPrice = orderPrice;
+//    }
+//
+//    public int getCount() {
+//        return count;
+//    }
+//
+//    public void setCount(int count) {
+//        this.count = count;
+//    }
+//}
